Start project restructuring
